agents/accounting/requirements.txt
pandas
pyyaml
reportlab
flask
gspread
oauth2client
python-telegram-bot
plaid-python
openpyxl
requests
